+++ src/GRA.Controllers/MappingProfile.cs
@@ -1,5 +1,6 @@
 ﻿using GRA.Controllers.ViewModel.Challenges;
 using GRA.Controllers.ViewModel.Join;
+using GRA.Controllers.ViewModel.MissionControl.Participants;
 using GRA.Domain.Model;
 
 namespace GRA.Controllers
@@ -13,6 +14,7 @@ namespace GRA.Controllers
             CreateMap<Step1ViewModel, User>().ReverseMap();
             CreateMap<Step2ViewModel, User>().ReverseMap();
             CreateMap<Step3ViewModel, User>().ReverseMap();
+            CreateMap<ParticipantsAddViewModel, User>().ReverseMap();
             CreateMap<TaskDetailViewModel, ChallengeTask>().ReverseMap();
         }
     }