+++ src/GRA.Controllers/MissionControl/ParticipantsController.cs
@@ -157,11 +157,212 @@ namespace GRA.Controllers.MissionControl
             }
             if (programId.HasValue)
             {
-                    viewModel.ProgramName =
-                        (await _siteService.GetProgramByIdAsync(programId.Value)).Name;
+                viewModel.ProgramName =
+                    (await _siteService.GetProgramByIdAsync(programId.Value)).Name;
+            }
+
+            var siteStage = GetSiteStage();
+            if (siteStage == SiteStage.RegistrationOpen || siteStage == SiteStage.ProgramOpen)
+            {
+                viewModel.CanSignUpParticipants = true;
             }
 
             return View(viewModel);
+        }
+
+        public async Task<IActionResult> Add()
+        {
+            PageTitle = "Add Participant";
+            var site = await GetCurrentSiteAsync();
+            var siteStage = GetSiteStage();
+            if (siteStage <= SiteStage.BeforeRegistration)
+            {
+                ShowAlertInfo("Registration has not opened yet");
+                return RedirectToAction("Index", "Participants");
+            }
+            else if (siteStage >= SiteStage.ProgramEnded)
+            {
+                ShowAlertInfo("The program has ended, participants cannot be added");
+                return RedirectToAction("Index", "Participants");
+            }
+
+            var systemList = await _siteService.GetSystemList();
+            var programList = await _siteService.GetProgramList();
+            var programViewObject = _mapper.Map<List<ProgramViewModel>>(programList);
+            var districtList = await _schoolService.GetDistrictsAsync();
+
+            ParticipantsAddViewModel viewModel = new ParticipantsAddViewModel()
+            {
+                RequirePostalCode = site.RequirePostalCode,
+                ProgramJson = Newtonsoft.Json.JsonConvert.SerializeObject(programViewObject),
+                SystemList = new SelectList(systemList.ToList(), "Id", "Name"),
+                ProgramList = new SelectList(programList.ToList(), "Id", "Name"),
+                SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name")
+            };
+
+            if (systemList.Count() == 1)
+            {
+                var systemId = systemList.SingleOrDefault().Id;
+                var branchList = await _siteService.GetBranches(systemId);
+                if (branchList.Count() > 1)
+                {
+                    branchList = branchList.Prepend(new Branch() { Id = -1 });
+                }
+                else
+                {
+                    viewModel.BranchId = branchList.SingleOrDefault().Id;
+                }
+                viewModel.BranchList = new SelectList(branchList.ToList(), "Id", "Name");
+                viewModel.SystemId = systemId;
+            }
+
+            if (programList.Count() == 1)
+            {
+                var programId = programList.SingleOrDefault().Id;
+                var program = await _siteService.GetProgramByIdAsync(programId);
+                viewModel.ProgramId = programList.SingleOrDefault().Id;
+                viewModel.ShowAge = program.AskAge;
+                viewModel.ShowSchool = program.AskSchool;
+            }
+
+            return View(viewModel);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> Add(ParticipantsAddViewModel model)
+        {
+            var site = await GetCurrentSiteAsync();
+            if (site.RequirePostalCode && string.IsNullOrWhiteSpace(model.PostalCode))
+            {
+                ModelState.AddModelError("PostalCode", "The Zip Code field is required.");
+            }
+
+            bool askAge = false;
+            bool askSchool = false;
+            if (model.ProgramId.HasValue)
+            {
+                var program = await _siteService.GetProgramByIdAsync(model.ProgramId.Value);
+                askAge = program.AskAge;
+                askSchool = program.AskSchool;
+                if (program.AgeRequired && !model.Age.HasValue)
+                {
+                    ModelState.AddModelError("Age", "The Age field is required.");
+                }
+                if (program.SchoolRequired)
+                {
+                    if (!model.NewEnteredSchool && !model.SchoolId.HasValue)
+                    {
+                        ModelState.AddModelError("SchoolId", "The School field is required.");
+                    }
+                    else if (model.NewEnteredSchool
+                        && string.IsNullOrWhiteSpace(model.EnteredSchoolName))
+                    {
+                        ModelState.AddModelError("EnteredSchoolName", "The School Name field is required.");
+                    }
+                }
+                if (model.NewEnteredSchool && !model.SchoolDistrictId.HasValue
+                    && ((program.AskSchool && !string.IsNullOrWhiteSpace(model.EnteredSchoolName))
+                        || program.SchoolRequired))
+                {
+                    ModelState.AddModelError("SchoolDistrictId", "The School District field is required.");
+                }
+            }
+
+            if (ModelState.IsValid)
+            {
+                if (!askAge)
+                {
+                    model.Age = null;
+                }
+                if (askSchool)
+                {
+                    if (model.NewEnteredSchool)
+                    {
+                        model.SchoolId = null;
+                    }
+                    else
+                    {
+                        model.EnteredSchoolName = null;
+                    }
+                }
+                else
+                {
+                    model.SchoolId = null;
+                    model.EnteredSchoolName = null;
+                }
+
+                User user = _mapper.Map<User>(model);
+                user.SiteId = site.Id;
+                try
+                {
+                    var newUser = await _userService.RegisterUserAsync(user, model.Password,
+                        model.SchoolDistrictId, true);
+                    if (UserHasPermission(Permission.EditParticipants))
+                    {
+                        return RedirectToAction("Detail", "Participants", new { id = newUser.Id });
+                    }
+                    else
+                    {
+                        return RedirectToAction("Index", "Participants");
+                    }
+                       
+                }
+                catch (GraException gex)
+                {
+                    ShowAlertDanger("Could not create participant account: ", gex);
+                    if (gex.Message.Contains("password"))
+                    {
+                        ModelState.AddModelError("Password", "Please correct the issues with the password.");
+                    }
+                }
+            }
+            PageTitle = "Add Participant";
+
+            if (model.SystemId.HasValue)
+            {
+                var branchList = await _siteService.GetBranches(model.SystemId.Value);
+                if (model.BranchId < 1)
+                {
+                    branchList = branchList.Prepend(new Branch() { Id = -1 });
+                }
+                model.BranchList = new SelectList(branchList.ToList(), "Id", "Name");
+            }
+            var systemList = await _siteService.GetSystemList();
+            var programList = await _siteService.GetProgramList();
+            var programViewObject = _mapper.Map<List<ProgramViewModel>>(programList);
+            model.SystemList = new SelectList(systemList.ToList(), "Id", "Name");
+            model.ProgramList = new SelectList(programList.ToList(), "Id", "Name");
+            model.ProgramJson = Newtonsoft.Json.JsonConvert.SerializeObject(programViewObject);
+            model.RequirePostalCode = site.RequirePostalCode;
+            model.ShowAge = askAge;
+            model.ShowSchool = askSchool;
+
+            var districtList = await _schoolService.GetDistrictsAsync();
+            if (model.SchoolId.HasValue)
+            {
+                var schoolDetails = await _schoolService.GetSchoolDetailsAsync(model.SchoolId.Value);
+                var typeList = await _schoolService.GetTypesAsync(schoolDetails.SchoolDisctrictId);
+                model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name",
+                    schoolDetails.SchoolDisctrictId);
+                model.SchoolTypeList = new SelectList(typeList.ToList(), "Id", "Name",
+                    schoolDetails.SchoolTypeId);
+                model.SchoolList = new SelectList(schoolDetails.Schools.ToList(), "Id", "Name");
+            }
+            else
+            {
+                model.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name");
+                if (model.SchoolDistrictId.HasValue)
+                {
+                    var typeList = await _schoolService.GetTypesAsync(model.SchoolDistrictId);
+                    model.SchoolTypeList = new SelectList(typeList.ToList(), "Id", "Name",
+                        model.SchoolTypeId);
+                    var schoolList = await _schoolService.GetSchoolsAsync(model.SchoolDistrictId,
+                        model.SchoolTypeId);
+                    model.SchoolList = new SelectList(schoolList.ToList(), "Id", "Name");
+                }
+            }
+
+            return View(model);
         }
 
         [Authorize(Policy = Policy.DeleteParticipants)]